requirements.txt
Django~=3.1.2
djangorestframework~=3.12.1
celery~=5.0.0
kavenegar~=1.1.2
requests~=2.22.0
google-auth-oauthlib~=0.4.1
google-api-python-client~=1.12.3
psycopg2~=2.8.6
djangorestframework-simplejwt~=4.4.0
django-celery-beat~=2.0.0
google-api-python-client~=1.12.3
dj-config-url~=0.1.1